LMS-1569 Moved methods from IFeatureVectorDAO to IImagingUploadDAO so that they share the same transaction context.
SVN: 16493

+++ screening/source/java/ch/systemsx/cisd/openbis/dss/etl/dataaccess/ImgFeatureValuesDTO.java
@@ -94,7 +94,7 @@ public class ImgFeatureValuesDTO extends AbstractHashable
         return featureDefId;
     }
 
-    public void setFeatureDefId(Integer featureDefId)
+    public void setFeatureDefId(long featureDefId)
     {
         this.featureDefId = featureDefId;
     }
